Add ability to emulate degraded network in Call via field trial
This is especially useful in Chrome, allowing use to emulate network
conditions in incoming or outgoing media without the need for platform
specific tools or hacks. It also doesn't interfere with the rest of the
network traffic.

Also includes some refactorings and a fix for crash on exit in loopback
tests.

Bug: None
Change-Id: I2656a2d4218acbe7f8ffd669de19a02275735438