Update patch set 3
Patch Set 3:

> Jonas, can you estimate the actual space saving benefit?
The measured difference is 4096 bytes, but that's smells like "a page". The true difference should be something like 60 bytes per log statement * 50 statements, so 3000 bytes.
 
> Also, note that the string operations used to be optional (no concats when logging not enabled) but now are not, this *may* have an effect on performance.
True, but if it only happens when video streams are removed that shouldn't make much of a difference.

Patch-set: 3
Reviewer: Gerrit User 6547 <6547@58829da1-049c-39fb-b951-ebdcd0984767>