NOTE: the changes in this CL are landed in [1] and reviewed in [2] due to some hiccup in the review system.
AGC2-fixed-digital: InterpolatedGainCurve with tests.

We map the signal level to the applied gain. This is a central part of
the FixedDigital mode of AGC2. The other central part is to compute
the signal level. That will be done in a later CL.

The mapping is done by
InterpolatedGainCurve::LookUpGainToApply(level). The gain curve
consists of 32 linear regions. The regions are parametrized as
"y=m*x+q". IGC contains 3 arrays X, M, Q. X is the boundaries of the
regions. M and Q are the parameters.

This CL also contains the code that computes the arrays. The computing
code is not compiled into libjingle_peerconnection.so. It's included
in the tests to compare the arrays to the computed values.

We first describe the gain curve in the DB scale. This is done by the
Limiter class. Then we use an algorithm (developed by alessiob@) to
select the best region boundaries and the optimal M and Q
coefficients.

[1] https://webrtc.googlesource.com/src/+/a05ee82c4ccb2a69dc726eb75c5868a9d1151aad
[2] https://webrtc-review.googlesource.com/c/src/+/54361
Bug: webrtc:7949
Change-Id: If67f92678d58f41983b7313f4920f511b393b42f