Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Not sure what these tests cover. Only RTP-related test seems to be the last, UpdateLifetime, which uses its own MockTransport.
> 
> Currently, RtpTransportControllerSend is more or less a collection of PacketRouter, Pacer and SendSideCongestionController. Using a fake implementation of RtpTransportControllerSendInterface would make sense in case we don't need to involve those classes in the tests, but I guess AudioSendStream still needs direct access via RtpTransportControllerSendInterface::send_side_cc() ?

Yes, but after a recent code change it doesn't actually access send_side_cc() directly, rather, the fake controller emulates the behavior of the real controller which leads to unfortunate duplicated implementations. By landing this CL it will be possible to remove the Fake in a later CL. In a later CL a Mock implementation of RtpTransportControllerSendInterface is introduced, that could be used to redesign these tests to use a complete mock.

Patch-set: 4
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>