Update patch set 5
Patch Set 5:

Thanks for a lot for the explanation.

> I think the goal would be to take Call out of the per-packet pipeline, and merge RtpStreamReceiverController with RtpTransport (or at least move ownership).

Taking Call out of the pipeline is great. I think I'd like to keep an interface representing the RtpTransport services for receive streams, to be passed to [Audio|Video]ReceiveStream constructors. That's currently RtpReceiveStreamcontroller. Besides interface to register RtpPacketSinks with the demuxer, it should have methods to emit RTCP messages, e.g, key frame requests, and possibly access to receive stats. I don't have any strong opinion on what it is called or the organization if the implementation, but I think we should have an interface like that and not give receive streams access to all of RtpTransport.

> The [Audio|Video]ReceiveStream classes could implement RtpPacketSinkInterface (or expose a "sink_interface()" accessor). Then BaseChannel/MediaChannel would be responsible for setting up demuxing, and also act as sinks for unsignaled SSRCs.

Note some streams have multiple sinks (more on this below). I kind of like the model that receive streams register their appropriate sinks them selves. E.g, VideoReceiveStream does not implement RtpPacketSinkInterface, but it creates one media sink (RtpVideoStreamReceiver) and one RTX sink (RtxReceiveStream), and registers them with the demuxer.

> So, the RTP packet pipeline would become:
> 
> IceTransport->
>  DtlsTransport->
>   RtpTransport->
>    RtpDemuxer->
>     MediaChannel (for unsignaled SSRCs)
>     AudioReceiveStream
>     VideoReceiveStream
> 
> And the configuration path would be:
> 
> PeerConnection::SetLocalDescription->
>  BaseChannel::SetLocalContent->
>   MediaChannel::AddRecvStream->
>    Call::Create[Audio|Video]RecvStream
>    RtpTransport::AddSink

Sounds like a pretty good plan, I only find the responsibility of AddSink somewhat worrying.
    
> The concrete steps to do this would be something like:
> 
> 1. Allow Call to take pre-parsed RTP packets (this CL).

Can you make Call implement RtpPacketSinkInterface in this cl? Or do you still need the return value for unsignalled ssrcs?

> 2. Use RtpDemuxer for the PT-based demuxing in BaseChannel, with one RtpDemuxer per BaseChannel.

What are your thoughts on payload_type_frequency? I think it's highly desirable to set that early and pass immutable RtpPacketReceived objects down the pipeline. And as far as I understand, setting up a mapping of payload_type to frequency should be doable up front, as soon as negotiation (or renegotiation) is finished. In the signalling path above, it could then be passed from PeerConnection to BaseChannel. 

> 3. Move RtpDemuxer ownership to RtpTransport; now BaseChannels will share an RtpDemuxer if bundled.

Each RtpTransport will still have only one RtpDemuxer, and there will be one RtpTransport per RTP session / ssrc space? (I'm afraid I still have a very fuzzy understanding of the responsibilities of BaseChannel and MediaChannel).

My plan has been to inject the RtpTransport into Call, and Call wouldn't do much with it besides passing it on to [Audio|Video]ReceiveStream constructors (srte@ currently has a series of cls doing that for the send side of things). But I'm afraid that organization won't scale that well to fully unbundled; Call::Create*ReceiveStream would need to to know which of a potential large set of Transports to use.

Maybe the transport should rather be part of VideoReceiveStream::Config, and leave Call unaware of RtpTransport?

> 4. Make MediaChannel responsible for setting up per-stream demuxing, exposing RtpPacketSinkInterface from receive stream classes, as described above.

This is the only point I find somewhat questionable. My model is that the receive stream constructor gets an RtpReceiveStreamController (or whatever that interface is replaced with), and register appropriate sinks in the demuxer. MediaChannel's role would only be to tell Call or the receive stream which transport to use.

This could be turned around (instead of having MediaChannel setting transport and ssrcs in the *ReceiveStream::Config, call the demuxer itself. Some difficulties with that is that (i) some receive streams register multiple sinks, e.g, media + rtx, (ii) the receive stream needs some access to the transport for other reasons, like sending RTCP messages. It's not entirely clear to me what you're suggesting. It might also be good to think about where responsibility for creating and registering FlexFEC streams should be. Should MediaChannel know the details of RTX and FlexFEC plumbing?

BTW, are you aware of Danil's work on RtcpTransciever? The idea is to have a single Rtcp object representing all Rtcp for one RTP session. Not yet used in webrtc, but I think RtpTransport should have one of those.

I

Patch-set: 5
Reviewer: Gerrit User 5234 <5234@58829da1-049c-39fb-b951-ebdcd0984767>