Update patch set 5
Patch Set 5:

> Taking Call out of the pipeline is great. I think I'd like to keep an interface representing the RtpTransport services for receive streams, to be passed to [Audio|Video]ReceiveStream constructors. That's currently RtpReceiveStreamcontroller. Besides interface to register RtpPacketSinks with the demuxer, it should have methods to emit RTCP messages, e.g, key frame requests, and possibly access to receive stats. I don't have any strong opinion on what it is called or the organization if the implementation, but I think we should have an interface like that and not give receive streams access to all of RtpTransport.
> 

That sounds good. So we'd probably have RtpTransport implement RtpReceiveStreamControllerInterface.

> Sounds like a pretty good plan, I only find the responsibility of AddSink somewhat worrying.

We can keep streams responsible for registering themselves as sinks then; it just may require them to know about metadata (such as MID) that they otherwise wouldn't need to know about.

> Can you make Call implement RtpPacketSinkInterface in this cl? Or do you still need the return value for unsignalled ssrcs?

It won't be a direct sink for the demuxer, and we do still need the return value for unsignalled SSRCs, so I don't think implementing RtpPacketSinkInterface would make sense.

> What are your thoughts on payload_type_frequency? I think it's highly desirable to set that early and pass immutable RtpPacketReceived objects down the pipeline. And as far as I understand, setting up a mapping of payload_type to frequency should be doable up front, as soon as negotiation (or renegotiation) is finished. In the signalling path above, it could then be passed from PeerConnection to BaseChannel.

That's definitely doable; the MediaChannel subclasses store all the codec information from SDP.

> Each RtpTransport will still have only one RtpDemuxer, and there will be one RtpTransport per RTP session / ssrc space? (I'm afraid I still have a very fuzzy understanding of the responsibilities of BaseChannel and MediaChannel).

Yes. And the pair of BaseChannel/MediaChannel corresponds to one m= section in SDP, which is why we need to worry about having multiple of them when supporting Unified Plan SDP.
 
> My plan has been to inject the RtpTransport into Call, and Call wouldn't do much with it besides passing it on to [Audio|Video]ReceiveStream constructors (srte@ currently has a series of cls doing that for the send side of things). But I'm afraid that organization won't scale that well to fully unbundled; Call::Create*ReceiveStream would need to to know which of a potential large set of Transports to use.
> 
> Maybe the transport should rather be part of VideoReceiveStream::Config, and leave Call unaware of RtpTransport?

Yes, that seems cleaner to me. 

> > 4. Make MediaChannel responsible for setting up per-stream demuxing, exposing RtpPacketSinkInterface from receive stream classes, as described above.
> 
> This is the only point I find somewhat questionable. My model is that the receive stream constructor gets an RtpReceiveStreamController (or whatever that interface is replaced with), and register appropriate sinks in the demuxer. MediaChannel's role would only be to tell Call or the receive stream which transport to use.

That sounds good to me; it's not that different than what I had in mind. The important part is moving the ownership of the RtpReceiveStreamController(s) out of Call.

> BTW, are you aware of Danil's work on RtcpTransciever? The idea is to have a single Rtcp object representing all Rtcp for one RTP session. Not yet used in webrtc, but I think RtpTransport should have one of those.

This is the first I've heard of it, but I agree, it would be appropriate to go in RtpTransport.



Anyway: As Zhi said, we're going to pursue another approach that won't require these temporary changes to call/ and media/. The downside is that we'll end up parsing the RTP packet twice, so we'll have see if that has a noticeable performance impact.

Patch-set: 5
Reviewer: Gerrit User 5053 <5053@58829da1-049c-39fb-b951-ebdcd0984767>