Update patch set 7
Patch Set 7:

Any updates here? Looks like the specification update PR on github will take a some time to be approved. Can we go forward with this CL to land it before M66 cut? Or maybe we should instead just add it to the old GetStats with googHugeFramesSent name and move it to the new stats collector after specification is updated?

Patch-set: 7
Reviewer: Gerrit User 5117 <5117@58829da1-049c-39fb-b951-ebdcd0984767>