Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> As per request from the 1st party, returned googHugeFramesSent to the old getstats. They don't parse it yet, and won't do it just for one metric. So, for now, this metric will be added in both collectors. I hope it's alright with you, Henrik, as it doesn't create any new work for migrating to rtcstatscollector, as metric is already added there.

The "let's put it in old stats for now" approach is bad, it increases the dependency on the non-standard API and postpone fixing the stats process to make standardizing a viable path forward. I'm postponing it myself due to lack of resources and lack of incentive to do so (e.g. little pressure from 1P to do so, and high priority items in focus unrelated to stats). And 1P postpones the standard process because they perceive it as too slow (and they're not wrong), which lessens the pressure on us. We're at a stalemate.

I need to formalize the stats process and pull some strings. The more we do this "put it in old stats" the deeper the hole we're digging for ourselves. I'd like to say no to adding stuff to old stats, a chess move to create that missing incentive. But I don't want to block 1P work before I've kept my end of the bargain and smoothened the process.

About this CL. Did you abandon it or do you want to land it for old stats? Because there are plenty of new stats changes still here.

Patch-set: 9
Reviewer: Gerrit User 5142 <5142@58829da1-049c-39fb-b951-ebdcd0984767>