Update patch set 2
Patch Set 2:

I looked into implementing this with a multimap, but that stores the values for a certain key according to some compare function. And we wouldn't really have any way to decide which decoder to use for a format.

I believe the map implementation and the DCHECK was originally just there to make sure users don't try to insert duplicate codecs. So instead of crashing on this, I now just check if the codec is already in the list and if so log a warning and move on to the next codec. What do you think?

Patch-set: 2
Reviewer: Gerrit User 5098 <5098@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5249 <5249@58829da1-049c-39fb-b951-ebdcd0984767>
Work-in-progress: false