Initializing Probe Controller in constructor.
This moves the initialization of the probe controller to the
constructor. This reduces the number of states it can be in.
Since the probe controller sends a probe as soon as it has an
initial configuration, this means it will now do this upon
construction. Previous changes guarantees that it's only constructed
when network is available, this means probing is safe and the
SetNetworkAvailable function is no longer needed.

Bug: webrtc:8415
Change-Id: Ica79bdea75d62be58fbe012f7438e5bd62e7a3d9