Introducing race checkers in video send stream.
The threading behavior of video send stream in needlessly restricted due
to usage of sequenced task checkers that checks for thread identity
rather than for actual races. This means that functions called at
construction time must be called on the right thread, forcing
construction to happen on the target thread as well. This complicates
the initialization behavior.

This change prepares for a later CL removing the posting of a
construction task to create VideoSendStreamImpl.

Bug: None
Change-Id: I7ce692a3083542490a939573decd1f4c4df6e86e