Removes construction of VideoSendStream on task queue.
This CL removes the construction of VideoSendStreamImpl by posting
a task on a task queue. This means that we no longer need to wait on
a posted task in the internal::VideoSendStream constructor.

It is not generally safe to wait on posted tasks as it can cause thread
exhaustion.

Introducing race checkers in video send stream.

The threading behavior of video send stream in needlessly restricted due
to usage of sequenced task checkers that checks for thread identity
rather than for actual races. This means that functions called at
construction time must be called on the right thread, forcing
construction to happen on the target thread as well. This complicates
the initialization behavior.

This change prepares for a later CL removing the posting of a
construction task to create VideoSendStreamImpl.

Change-Id: I044435716c1e44363ba36ecce9eb716fe0fc260b