Replaced DestructAndGetRtpStateTask with lambda.
Slight change in functionality: send_stream_ member is no longer moved
to the QueuedTask. This means that a possible race on access to
send_stream_ will not cause nullpointer dereferencing until the posted
task has been run. Most usages of send_stream_ are protected by
thread_checker_, but not DeliverRtcp and EnableEncodedFrameRecording.

Bug: None
Change-Id: Ife42071a4aa2811fcaf2f3ef21ca1888e6640ca3