Update patch set 1
Patch Set 1: Code-Review+1

lgtm.

Although the class is not a template right now, it has a member variable and ctor that effectively make it a template. In order to actually make it not a template, aside from the ctor, will we change the member variable to store begin_, end_ (or size) from the ArrayView variable?

Patch-set: 1
Reviewer: Gerrit User 5508 <5508@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1