Update patch set 5
Patch Set 5:

(2 comments)

The death tests didn't work out of the box, because the code failed to DCHECK-die in all cases when the buffer was too small. I fixed that, but that means this CL is now quite a bit larger, and Tommi's LGTM can no longer be considered valid. So please review again...

Patch-set: 5
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>