Removing synchronous call on transport controller.
Removing the Synchronous call AvailableBandwidth from the
RtpTransportControllerSend interface. This prepares for removing
locking behavior in SendSideCongestionController that exists just
to support this feature.

To keep backwards compatibility with the old SSCC, the struct
TargetTransferRate is constructed in RtpTransportControllerSend.
This step can be removed in the future when the old SSCC is
deprecated.

Bug: webrtc:8415
Change-Id: I06f64a89848157de412901c989650d1ecf35246b