Update patch set 3
Patch Set 3:

There is no condition that turns transport feedback availability off, so if we have ever seen a packet with transport feedback on any stream we'll assume that we have it forever. This might be fine, but I'd like to discuss it with stefan first.

Patch-set: 3
Reviewer: Gerrit User 5150 <5150@58829da1-049c-39fb-b951-ebdcd0984767>