Add internal event structure and stats generation to IceEventLog.
The current IceEventLog serves as an interface to generate ICE events by
the RtcEventLog convention and relays the events to the underlying
RtcEventLog infrastructure. Useful stats from the network stack is not
directly accessible via the stats collection mechanism. IceEventLog is
extended with internal event structure to allow event buffering and
stats generation.

Bug: webrtc:9031
Change-Id: I1ce0377c4982905f5f6e9948bd18b30caa17a90c