Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> unit test, if this func is important?


Currently VideoSendStreamImpl doesn't have unit tests. (internal::VideoSendStream has call tests though) It might be good to put it in it's own file and add specific unit tests, but not sure if it's worth to do it just for this feature. 

I'm thinking that for this functionality it might make more sense to have an end to end test for the complete functionality, to ensure that the full chain works. WDYT?

Patch-set: 1
Reviewer: Gerrit User 5531 <5531@58829da1-049c-39fb-b951-ebdcd0984767>