Update patch set 8
Patch Set 8:

I checked and in fact you are right 'RegisterRefreshAndMoveHandlers' and the stream handler are called in the same thread. (because of CFRunLoopAddSource(CFRunLoopGetCurrent()) ).
But 'CaptureFrame' is called in another thread.

This because chromium::content::DesktopCaptureDevice creates a dedicated capture thread. The webrtc::DesktopCapturer is created in the main thread (same for the 'Init' function) but 'Start', 'CaptureFrame' and the destructor are called from this dedicated thread.

This explain why I need the new lock in ScreenCapturerMacIOSurface. Also if you check "ScreenCapturerHelper helper_;" all its functions are protected with a lock.

Though according to the code it seems that 'RegisterRefreshAndMoveHandlers' is called from 'CaptureFrame' when the screen mode change on the fly and so it changes the logic above, and so current upstream implementation is inconsistent.
To make it consistent I think it should call 'Init' from the dedicated thread but that is probably something to tackle in a separate issue.

Patch-set: 8
Reviewer: Gerrit User 6236 <6236@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0