Update patch set 2
Patch Set 2: Code-Review+1

(3 comments)

The difference between the 0x4000-0x7FFF range and the 0xC000-0xFFFF range is that the former is "comprehension-required", meaning that an error should be returned if it isn't understood. Which it turns out we aren't honoring; see crbug.com/webrtc/9063

Just to confirm, do you want this "comprehension-required" behavior? Will a WebRTC endpoint ever receive the custom attribute you plan to define, or only send it? Because if you plan for a WebRTC endpoint to receive it, we may run into a snag when the bug linked above is fixed.

Patch-set: 2
Reviewer: Gerrit User 5053 <5053@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1