Update patch set 1
Patch Set 1: Code-Review+1

I think skipping initial frames is a good idea. While 60 is probably fine, I'm wondering if we reduce the number of skipped frames (e.g., to 40) to ensure we get enough data for shorter sequences.

Finally, and not directly related to this patch, it would be nice to tune the initial frames to mitigate the lousy quality at the beginning of the sequence.

Patch-set: 1
Reviewer: Gerrit User 6680 <6680@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Code-Review=+1