Update patch set 17
Patch Set 17:

(1 comment)

> Patch Set 16:
> 
> (1 comment)
Hi Zijie, thx for the suggestions and please take a look to Patch Set 17. In the case everything is good to go please hold on for an eventual merge as I do not have access to a secondary display this week. Thx!

Patch-set: 17
Reviewer: Gerrit User 6236 <6236@58829da1-049c-39fb-b951-ebdcd0984767>