Update patch set 2
Patch Set 2:

+Mirko as FYI

This is somewhat of an experiment.

I measured the binary size difference before and after Jonas' cl
https://webrtc-review.googlesource.com/c/src/+/64881

The difference was +16 bytes, which is perhaps not unexpected since stringstream is still being used widely. What's interesting here though, is that the difference was 16 bytes, not 4K.

So, I took a look at the AudioSendStream::Config class in the code size tool. The size of the methods were there reported to be 480+288+272=1040 bytes.
As it turns out, what keeps these methods from being optimized out, is this single RTC_LOG() method that I'm changing to be an RTC_DLOG. Once I did that, the binary size went down by 4K... the plot thickens...

Patch-set: 2
CC: Gerrit User 5087 <5087@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Gerrit User 5508 <5508@58829da1-049c-39fb-b951-ebdcd0984767>