Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> +ossu - can you comment on the importance of this log line for non-dcheck builds?

I think it's reasonably important as far as logging lines go: it does show all of the settings that goes into the AudioSendStream. Not sure why it would have to take 4k (or even 1k) of code to write that all, though.

Also, I haven't actually used that actual information in any of my bug hunts yet, but I can see it coming in handy; i.e. to confirm that things are actually being configured the way we expect them to.

Would it be possible to try and optimize the log instead (for now). I see it's at least creating two stringstreams - one in Config::ToString and one in Config::Rtp::ToString. Merge those and switch to snprintf, perhaps?

Patch-set: 2
Reviewer: Gerrit User 5399 <5399@58829da1-049c-39fb-b951-ebdcd0984767>