Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> I'm not convinced this mechanism is worth the new concept and added code in the templates. I consider adding codes to templates to be costly, because they 1) are very central to the build system and 2) anything added there causes us to deviate from Chromium.
> 
> It seems to me we can get almost all the benefit by this change by adding an assert_no_deps to libjingle_peerconnection_datachannelonly_so, at least if the structure things appropriately by dir. assert_no_deps = ["../../audio_codecs/*"] would be better than assert_no_deps = ["../../audio_codecs:audio_decoder_opus/", "../../audio_codecs:audio_decoder_opus", ...] for instance, but I guess we can't do the former because for instance api/audio_codecs:audio_codecs_api still needs to be visible. 
> 
> Am I missing a bigger picture here? Are there targets other than libjingle_peerconnection_datachannelonly_so and maybe peerconnectionfactory_no_media_objc that would benefit from this protection? It seems to me it would just be annoying if someone refactored in a new target between audio codecs and libjingle_peerconnection, for instance.
> 
> There are downsides with just using assert_no_deps, though, but I think they're small. It will detect if blacklisted targets are moved or renamed (the label becomes invalid), but it fails to detect if new codecs are added. I don't think that's a big problem.
> 
> I'm not convinced, therefore, at least not yet.

Just libjingle_peerconnection_datachannelonly_so and peerconnectionfactory_no_media_objc doesn't seem like it would be enough. For the case of audio codecs, for example, we'd ideally want to tag *every* user-visible build target except the audio codecs/factories themselves with assert_no_deps = audio_codec_targets. This is obviously infeasible to do by hand, so we'd have to limit ourselves to the ones that are likely to catch unwanted audio codec dependencies. But which ones are that? (Not a rhetorical question---I'm not familiar enough with our high-level build targets.)

The advantage with the "poison" system is that you only have to tag the dangerous stuff and the targets that *are* allowed to depend on it, and *not* every target that *isn't* allowed to depend on it. It's not really more powerful than assert_no_deps, just more convenient and therefore easier to get right. Just like with `testonly`, where you tag the test-only targets, and all the other targets by default aren't test-only and therefore aren't allowed to depend on test-only targets. It would be kind of annoying if you had to tag all your top-level targets as "non-test" in order to get errors if they depended on test-only code.

I'll try using assert_no_deps directly to get a better sense of how it works in practice.

Patch-set: 9
Reviewer: Gerrit User 5126 <5126@58829da1-049c-39fb-b951-ebdcd0984767>
Label: Commit-Queue=0